Fix some warnings

+++ lib/database/src/so/elements/SoAccumulatedElement.cpp
@@ -276,7 +276,7 @@ SoAccumulatedElement::print(FILE *fp) const
 
     fprintf(fp, "NodeIds: [ ");
     for (int i = 0; i < nodeIds.getLength(); i++) {
-	fprintf(fp, "%d, ", nodeIds[i]);
+    fprintf(fp, "%ld, ", (unsigned long)nodeIds[i]);
     }
     fprintf(fp, "]\n");
 }
